slett ekstra løserne
