+++ helsebro/src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/helsebro/App.kt
@@ -21,8 +21,8 @@ fun main() {
 
 fun RapidsConnection.createHelsebro(priProducer: PriProducer): RapidsConnection {
     TrengerForespoerselLøser(this, priProducer)
-    TrengerForespoerselLøser2(this, priProducer)
+    TrengerForespoerselLøser(this, priProducer)
     ForespoerselSvarLøser(this)
-    ForespoerselSvarLøser2(this)
+    ForespoerselSvarLøser(this)
     return this
 }
